requirements.txt
langchain-core
langchain-community
sentence-transformers
chromadb
transformers
llama-cpp-python
langchain-chroma
langchain-huggingface
tokenizers
fastapi
uvicorn
